% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_boundaries}
\alias{county_boundaries}
\title{Kenya County Spatial Boundaries}
\format{
An sf data frame with variables:
\describe{
  \item{county_code}{Numeric county code (1-47)}
  \item{county_name}{Name of the county}
  \item{area_km2}{Area of county in square kilometers}
  \item{geometry}{Spatial geometry (polygon)}
}
}
\source{
Kenya National Bureau of Statistics, Survey of Kenya
}
\usage{
county_boundaries
}
\description{
Spatial polygons (sf object) representing the boundaries of Kenya's 47 counties.
}
\keyword{datasets}
