% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_info}
\alias{county_info}
\title{Kenya County Administrative Information}
\format{
A data frame with variables:
\describe{
  \item{county_code}{Numeric county code (1-47)}
  \item{county_name}{Name of the county}
  \item{county_capital}{County capital/headquarters}
  \item{established_date}{Date county government was established}
}
}
\source{
Kenya National Bureau of Statistics
}
\usage{
county_info
}
\description{
Administrative and demographic information for Kenya's counties.
}
\keyword{datasets}
