% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{county_stats}
\alias{county_stats}
\title{Calculate County Statistics}
\usage{
county_stats()
}
\value{
A data frame with county statistics
}
\description{
Get summary statistics about counties (area, population, etc.).
}
\examples{
county_stats()
}
