% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countyGAV}
\alias{countyGAV}
\title{Kenya County Gross Value Added Data}
\format{
A data frame with variables:
\describe{
  \item{county_code}{Numeric county code (1-47)}
  \item{county_name}{Name of the county}
  \item{2019, 2020, 2021, 2022, 2023}{GAV values by year}
}
}
\source{
Kenya National Bureau of Statistics
}
\usage{
countyGAV
}
\description{
County contribution to Gross Value Added for Kenya's counties, 2019-2023.
}
\keyword{datasets}
