% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countyGCP}
\alias{countyGCP}
\title{Kenya Gross County Product Data}
\format{
A data frame with economic indicators by year
}
\source{
Kenya National Bureau of Statistics
}
\usage{
countyGCP
}
\description{
Economic data showing Gross County Product for Kenya's counties by year.
}
\keyword{datasets}
