% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countyGovernance}
\alias{countyGovernance}
\title{Kenya County Governance Data}
\format{
A data frame with county governance data
}
\source{
Kenya Electoral Commission
}
\usage{
countyGovernance
}
\description{
Basic county governance information including governors and political parties.
}
\keyword{datasets}
