% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countyPopulation}
\alias{countyPopulation}
\title{Kenya County Population Data}
\format{
A data frame with variables:
\describe{
  \item{county_code}{Numeric county code (1-47)}
  \item{county_name}{Name of the county}
  \item{2019, 2020, 2021, 2022, 2023}{Population by year}
}
}
\source{
Kenya National Bureau of Statistics, Census data
}
\usage{
countyPopulation
}
\description{
Population statistics for Kenya's counties with yearly data.
}
\keyword{datasets}
