% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_governor}
\alias{get_governor}
\title{Get Governor Information by County}
\usage{
get_governor(county_name = NULL, county_code = NULL)
}
\arguments{
\item{county_name}{Character. Name of the county (optional)}

\item{county_code}{Numeric. County code 1-47 (optional)}
}
\value{
A data frame with governor information
}
\description{
Retrieve governor information for a specific county or all counties.
}
\examples{
get_governor("Nairobi")
get_governor(county_code = 47)
}
