% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{governors}
\alias{governors}
\title{Kenya County Governors Data}
\format{
A data frame with variables:
\describe{
  \item{county_code}{Numeric county code (1-47)}
  \item{county_name}{Name of the county}
  \item{governor_name}{Full name of the governor (2017-2022 term)}
  \item{political_party}{Political party of the governor}
  \item{deputy_governor_name}{Full name of the deputy governor}
  \item{term_start}{Start date of term (YYYY-MM-DD)}
  \item{term_end}{End date of term (YYYY-MM-DD)}
  \item{previous_governor}{Name of previous governor (if applicable)}
  \item{election_year}{Year of election}
}
}
\source{
Kenya Electoral Commission, County Government websites
}
\usage{
governors
}
\description{
A dataset containing information about current and former governors of Kenya's 47 counties.
}
\keyword{datasets}
