% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countyGCP}
\alias{countyGCP}
\title{Kenya Gross County Product Data}
\format{
A data frame with variables:
\describe{
  \item{county_code}{Numeric county code (1-47)}
  \item{county_name}{Name of the county}
  \item{2013, 2014, 2015, ...}{GCP values by year (cleaned from X2013 format)}
}
}
\source{
Kenya National Bureau of Statistics
}
\usage{
countyGCP
}
\description{
Economic data showing Gross County Product for Kenya's counties by year.
}
\keyword{datasets}
