% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countyPopulation}
\alias{countyPopulation}
\title{Kenya County Population Data}
\format{
A data frame with population data
}
\source{
Kenya National Bureau of Statistics, Census data
}
\usage{
countyPopulation
}
\description{
Population statistics for Kenya's counties.
}
\keyword{datasets}
